requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
Jinja2==3.0.3
itsdangerous==2.0.1
MarkupSafe==2.1.1
pyyaml==6.0.1
Flask-WTF==1.0.1
WTForms==3.0.1
python-dateutil==2.8.2
click>=8.0.0

